Warning.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@46916 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/opt.cpp
@@ -246,7 +246,7 @@ bool OPT_access_path(const jrd_req* request, UCHAR* buffer, SLONG buffer_length,
 
 	const size_t length = infoBuffer.getCount();
 
-	if (length > buffer_length)
+	if (length > static_cast<ULONG>(buffer_length))
 	{
 		*return_length = 0;
 		return false;
